+++ packages/core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # ascii-kit
 
+## 0.0.6
+
+### Patch Changes
+
+- add documentation to packages
+
+- Updated dependencies []:
+  - @ascii-kit/image@0.0.6
+  - @ascii-kit/font@0.0.6
+  - @ascii-kit/tree@0.0.6
+  - @ascii-kit/svg@0.0.6
+  - @ascii-kit/qr@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes