+++ packages/tree/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ascii-kit/tree
 
+## 0.0.6
+
+### Patch Changes
+
+- add documentation to packages
+
 ## 0.0.5
 
 ### Patch Changes